+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-ui
 
+## 2.8.57
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.57
+
+## 2.8.56
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.56
+
 ## 2.8.55
 
 ### Patch Changes